feat: 리뷰 API 검증 로직 추가 #4 - 리뷰 수정 및 삭제 API: 존재하는 리뷰인지, 요청한 user가 리뷰의 작성자인지 검증하는 코드 추가 - Javadoc 추가

+++ src/main/java/com/example/demo/domain/review/web/ReviewController.java
@@ -21,53 +21,89 @@ public class ReviewController {
     private final ReviewService reviewService;
 
     /**
-     * 리뷰 등록 API
+     * 리뷰 작성 API
+     * @param WriteReviewRequest request
+     * @return ApiResponse
      */
     @PostMapping
     public ApiResponse writeReview(@RequestBody WriteReviewRequest request) {
-
-        // 헤더로 받은 jwt 토큰을 복호화 해서 유저 정보 획득, 아마도 userId가 아닌 user 객체 자체를 변수로 사용할듯
-        // 더미 데이터
+        // 헤더로 받은 jwt 토큰을 복호화 해서 유저 정보 획득, 아마도 userId가 아닌 User 객체 자체를 변수로 사용할듯
+        // 현재는 더미 데이터 사용
         Long findUserId = 1L;
 
         // 리뷰 생성 및 등록
         Review review = Review.createReview(findUserId, request.getPlaceId(), request.getScore(), request.getContent(), request.getImgUrl(), request.getTendency());
         reviewService.writeReview(review);
-
         return new ApiResponse<>(200, "API OK");
     }
 
     /**
      * 리뷰 수정 API
+     * @param Long reviewId
+     * @param ModifyReviewRequest request
+     * @return ApiResponse
      */
     @PostMapping("/{reviewId}")
     public ApiResponse modifyReview(
             @PathVariable("reviewId") Long reviewId,
             @RequestBody ModifyReviewRequest request
     ) {
-        // reviewId로 review 찾아서 업데이트
+        // reviewId를 통해 review 조회
+        Review review = reviewService.findByReviewId(reviewId);
+
+        // review의 작성자가 맞는지 확인
+        Long userId = 1L;  // 더미 데이터
+        if (review.getUserId() != userId) {
+            throw new IllegalArgumentException("해당 기능에 대한 권한이 존재하지 않습니다.");
+        }
+
+        // 위 검증을 통과한다면, review를 업데이트
         reviewService.updateReview(reviewId, request);
         return new ApiResponse<>(200, "API OK");
     }
 
     /**
      * 리뷰 삭제 API
+     * @param Long reviewId
+     * @return ApiResponse
      */
     @DeleteMapping("/{reviewId}")
     public ApiResponse deleteReview(@PathVariable("reviewId") Long reviewId) {
+        // reviewId를 통해 review 조회
+        Review review = reviewService.findByReviewId(reviewId);
+
+        // review의 작성자가 맞는지 확인
+        Long userId = 1L;  // 더미 데이터
+        if (review.getUserId() != userId) {
+            throw new IllegalArgumentException("해당 기능에 대한 권한이 존재하지 않습니다.");
+        }
+
+        // 위 검증을 통과한다면, review를 삭제
         reviewService.deleteReview(reviewId);
         return new ApiResponse<>(200, "API OK");
     }
 
     /**
      * 리뷰 조회 API
+     * @param Long placeId
+     * @return ApiResponse<GetReviewsResponse>
+     * 이슈사항 : 삭제 시, deletedAt이 요청한 시간으로 업데이트 되는 형식 -> Status 사용하는 것이 더 좋아 보임
      */
     @GetMapping("/{placeId}")
     public ApiResponse<GetReviewsResponse> getReviews(@PathVariable("placeId") Long placeId) {
         // placeId로 해당 장소에 대한 리뷰 리스트 조회
-        List<Review> findReviews = reviewService.findReviewByPlaceId(placeId);
+        List<Review> findReviews = reviewService.findByPlaceId(placeId);
 
         GetReviewsResponse response = new GetReviewsResponse(findReviews);
-        return new ApiResponse<>(200, "API OK", response);
+        return new ApiResponse<GetReviewsResponse>(200, "API OK", response);
+    }
+
+    /**
+     * 연결 테스트 API
+     * 추후에 삭제할 예정입니다.
+     */
+    @GetMapping("/test")
+    public ApiResponse testApi() {
+        return new ApiResponse(200, "Test OK");
     }
 }
